refactor: migrate platform and reorganize project structure
Major refactor involving platform migration from Fly.io to Railway.app,
type system reorganization, and build improvements:

- Platform migration: Replace Fly.io config with Railway.app deployment
- Type system: Move types from src/lib/types/ to root types/ directory
- Pricing config: Move from centralized config to component-specific location
- Build: Add vite-tsconfig-paths for better TypeScript path resolution
- Cleanup: Remove unused documentation and changelog files
