--- CHANGELOG.md
@@ -1,7 +1,0 @@
-# @nextnode/nextnode-front
-
-## 1.0.0
-
-### Major Changes
-
-- [#1](https://github.com/NextNodeSolutions/nextnode-front/pull/1) [`be8321a`](https://github.com/NextNodeSolutions/nextnode-front/commit/be8321aae2cd9f065c5777732549d7b84b9ec9a9) Thanks [@walid-mos](https://github.com/walid-mos)! - Imported old-nextnode to this project.
