+++ package.json
@@ -68,6 +68,7 @@
     "prettier-plugin-astro": "^0.14.1",
     "prettier-plugin-tailwindcss": "^0.6.14",
     "typescript": "^5.0.0",
+    "vite-tsconfig-paths": "^5.1.4",
     "vitest": "^3.1.4"
   },
   "peerDependencies": {
