+++ railway.toml
@@ -1,0 +1,30 @@
+# Railway.toml Configuration
+# NextNode Front - Astro SSR Application with Docker
+
+[build]
+builder = "DOCKERFILE"
+dockerfilePath = "Dockerfile"
+watchPatterns = [
+    "src/**",
+    "public/**", 
+    "config/**",
+    "types/**",
+    "astro.config.mjs",
+    "package.json",
+    "pnpm-lock.yaml",
+    "tsconfig.json",
+    "Dockerfile",
+    "railway.toml"
+]
+
+[deploy]
+# Health check endpoint - Railway will use the Dockerfile's HEALTHCHECK
+healthcheckPath = "/"
+healthcheckTimeout = 30
+restartPolicyType = "ON_FAILURE"
+restartPolicyMaxRetries = 3
+
+# Environment-specific configurations
+[environments.production.deploy]
+healthcheckTimeout = 45
+restartPolicyMaxRetries = 5
