+++ src/components/features/pricing/PricingCard.astro
@@ -11,7 +11,7 @@ import {
 	PRICING_CARD_CLASSES,
 	getPlanCSSVars,
 } from '@/lib/ui/pricing-constants'
-import type { Plan } from '@/lib/types/plans'
+import type { Plan } from '@/types/plans'
 
 export interface Props {
 	plan: Plan
